refactor(d19): promote place_card test helper into share_block_helpers
